+++ Sources/BraintreeUIComponents/BTPayPalButtons.swift
@@ -1,25 +1,46 @@
 import SwiftUI
 
-struct PayPalButton: View {
+/// PayPal payment button. Available in the colors primary (PayPal blue), black, and white.
+public struct PayPalButton: View {
 
-    let type: BTPaymentButtonStyle
+    /// The style of the PayPal payment button. Available in the colors primary (PayPal blue), black, and white.
+    let style: BTPaymentButtonStyle
+
+    /// The width of the PayPal payment button. Minimum width is 131 points. Maximum width is 300 points.
+    let width: CGFloat?
+
+    /// The PayPal payment button action.
     let action: () -> Void
 
-    var body: some View {
+    public init(style: BTPaymentButtonStyle, width: CGFloat? = nil, action: @escaping () -> Void) {
+        self.style = style
+        self.width = width
+        self.action = action
+    }
+
+    public var body: some View {
         Button(action: action) {
             HStack {
-                Image(type.logoImage)
-                    .resizable()
-                    .scaledToFit()
-                    .frame(height: 26)
+                if let logoImage = style.logoImage {
+                    Image(logoImage)
+                        .resizable()
+                        .scaledToFit()
+                        .frame(height: 26)
+                }
             }
-            .frame(width: 300, height: 45)
-            .background(type.backgroundColor)
+            .frame(minWidth: 131, maxWidth: width ?? 300)
+            .frame(height: 45)
+            .background(style.backgroundColor)
             .cornerRadius(4)
             .overlay(
                 RoundedRectangle(cornerRadius: 4)
-                    .stroke(type.hasOutline ? Color.black : Color.clear, lineWidth: type.hasOutline ? 1 : 0)
+                    .stroke(
+                        style.hasOutline ? .black : .clear,
+                        lineWidth: style.hasOutline ? 1 : 0
+                    )
             )
+            .accessibilityLabel("PayPal checkout button")
+            .accessibilityHint("PayPal checkout button")
         }
     }
 }
@@ -29,13 +50,13 @@ struct ContentView: View {
     var body: some View {
         VStack(spacing: 16) {
             // Primary Button
-            PayPalButton(type: .primaryPayPal) {}
+            PayPalButton(style: .primaryPayPal, width: 300) {}
 
             // Black Button
-            PayPalButton(type: .black) {}
+            PayPalButton(style: .black, width: 250) {}
 
             // White Button
-            PayPalButton(type: .white) {}
+            PayPalButton(style: .white, width: 100) {}
         }
     }
 }